+++ packages/suite/src/support/extraDependencies.ts
@@ -8,11 +8,11 @@ import { closeModal, openModal } from '@suite/modal';
 import { createElectronPlatformEncryption } from '@suite/platform-encryption-electron';
 import { createWebauthnPlatformEncryption } from '@suite/platform-encryption-webauthn';
 import {
-    ensureRouterPath,
-    getPrefixedURL,
+    HistoryDep,
+    SuiteRouterHistoryDep,
+    createSuiteRouterHistory,
     selectRoute,
     selectRouterApp,
-    stripPrefixedURL,
 } from '@suite/router';
 import {
     DisableLegacyMetadataIfNeededDep,
@@ -53,12 +53,6 @@ import { selectIsWindowVisible } from 'src/reducers/suite/windowReducer';
 import { reportSecurityCheck } from 'src/utils/suite/sentry';
 import { fixLoadedCoinjoinAccount } from 'src/utils/wallet/coinjoinUtils';
 
-import {
-    HistoryDep,
-    SuiteRouterHistory,
-    SuiteRouterHistoryDep,
-    SuiteRouterHistoryDeps,
-} from './suite/suiteRouterHistory';
 import { forgetBluetoothDeviceThunk } from '../actions/bluetooth/bluetoothEraseBondsThunk';
 import { createDisableLegacyMetadataIfNeeded } from '../actions/suiteSync/disableLegacyMetadateIfNeeded';
 import type { BioAuthState } from '../reducers/bioAuth';
@@ -76,25 +70,6 @@ const connectInitSettings: ConnectInitSettings = {
     enableFirmwareHashCheck: true,
     firmwareHashCheckTimeouts: FW_HASH_CHECK_DEFAULT_TIMEOUTS,
 };
-
-export const createSuiteRouterHistory = ({
-    history,
-}: SuiteRouterHistoryDeps): SuiteRouterHistory => ({
-    getLocation: () => {
-        const { location } = history;
-
-        return ensureRouterPath({ ...location, pathname: stripPrefixedURL(location.pathname) });
-    },
-    navigate: (to, state) =>
-        history.push(
-            { ...to, pathname: to.pathname ? getPrefixedURL(to.pathname) : undefined },
-            state,
-        ),
-    listen: listener =>
-        history.listen(({ location, action }) =>
-            listener({ location: ensureRouterPath(location), action }),
-        ),
-});
 
 export type StoreAPIDep = {
     getState: () => any;
