+++ suite/router/src/index.ts
@@ -1,7 +1,9 @@
 export * from './anchors';
 export * from './anchorUtils';
 export * from './router';
+export * from './createSuiteRouterHistory';
 export * from './routerMiddleware';
 export * from './routerParams';
 export * from './routerReducer';
 export * from './routes';
+export * from './suiteRouterHistory';
