+++ suite/test-utils/src/initStoreForTests.ts
@@ -1,9 +1,7 @@
 import { createMemoryHistory } from 'history';
 
+import { createSuiteRouterHistory } from '@suite/router';
 import { PreloadedState, Store, initStore } from '@trezor/suite';
-// TODO fix deep import
-// eslint-disable-next-line local-rules/no-package-deep-imports
-import { createSuiteRouterHistory } from '@trezor/suite/src/support/extraDependencies';
 
 /**
  * Test-friendly wrapper for initStore that provides necessary dependencies like history.
