+++ suite/trading/tsconfig.json
@@ -8,9 +8,11 @@
     },
     "include": ["src/**/*", "styled.d.ts"],
     "references": [
-        { "path": "../../suite-common/geolocation" },
-        { "path": "../../suite/router" },
+        {
+            "path": "../../suite-common/geolocation"
+        },
         { "path": "../../suite-common/icons" },
+        { "path": "../router" },
         { "path": "../../suite-common/trading" },
         { "path": "../intl" },
         { "path": "../../packages/components" },
